Decisions and actions
Instead of `getAppByName` (which, by the name you can figure out how temporary it was), require a 'decide' function (even a worse name, but looks nice).

The `decide` function has the following signature: `function(req, name callback)` where `callback` is `function(err, decision)`. The decision has the following format:

```js
{
    // indicates if this app requires client authentication
    public: true | false,

    // indicates if this app will have https redirection on http calls
    secure: true | false,

    // the action to perform for this app
    action: { ... }
}
```

The `action` property may contain a single key that determines the "result" of the routing decision. A routing action is a `function(params, req, res, next)` where `params` are parameters returned by the `decide()` function and `req`, `res` and `next` are middleware functions. `req` and `res` are express.js request and response objects overloaded with some farmjs goodies:

 * __req.reqid__ - The uuid of the request, allocated per request and also added in the response header `x-farmjs-reqid`
 * __req.logger__ - A console-like contextual logger. Use it to emit logs in the context of handling this request
 * __req.secure__ - `true` if this is an HTTPS request
 * __req.hints__ - A hash of '$' hints passed in the query string.
 * __res.error(status, obj)__ - Will end the request with `status` HTTP status and `obj` as JSON output.
 * __res.proxy(port, host)__ - Will proxy the request to host:port and end it. If you wish to add headers to the _response_, emit them before calling `res.proxy()`

Action handlers can be installed using: `router.install(type, handler)` where `type` is the key in the `action` property of the decision.

The router is automatically installed with the following actions.

Spawns a child process with env.PORT allocation and proxy request to it

```js
action: {
    spawn: {
        // program to execute (defaults to node)
        command: process.execPath

        // environment variables to bind to the child process
        // farmjs will automatically add FARMJS_APP and FARMJS_INSTANCE
        env: { },

        // working directory for the child process
        cwd: null,

        // any other options passed to [spinner](https://github.com/anodejs/node-spinner)
        ...
    }
}
```

Proxy the request to some host:port

```js
action: {
    proxy: {
        host: 'localhost',
        port: ...,
        headers: { }
    }
}
```

End the request with an HTTP status code

```js
action: {
    error: {
        status: ..., // http status code

        // any other fields will be emitted in the application/json response
    }
}
```

```js
action: {
    redirect: {
        location: ..., // href to redirect to
    }
}
```

This action will cause farmjs to re-issue a `decide()` call with the `target` as the name of the name
of the app being looked up. This essentially allows easily aliasing apps be just emitting an alias action
to the target app.

```js
action: {
    alias: {
        target: ..., // name of app to alias
    }
}